+++ src/DataService/DataObjectFactory.php
@@ -170,13 +170,13 @@ class DataObjectFactory
     /**
      * Query for MPX data using with parameters.
      *
-     * @param ObjectListQuery $objectListQuery The fields and values to filter by. Note these are exact matches.
+     * @param ObjectListQuery $objectListQuery (optional) The fields and values to filter by. Note these are exact matches.
      * @param IdInterface     $account         (optional) The account context to use in the request. Defaults to the account
      *                                         associated with the authenticated client.
      *
      * @return ObjectListIterator An iterator over the full result set.
      */
-    public function select(ObjectListQuery $objectListQuery, IdInterface $account = null): ObjectListIterator
+    public function select(ObjectListQuery $objectListQuery = null, IdInterface $account = null): ObjectListIterator
     {
         return new ObjectListIterator($this->selectRequest($objectListQuery, $account));
     }
@@ -186,14 +186,18 @@ class DataObjectFactory
      *
      * @see \Lullabot\Mpx\DataService\DataObjectFactory::select
      *
-     * @param ObjectListQuery $objectListQuery The fields and values to filter by. Note these are exact matches.
+     * @param ObjectListQuery $objectListQuery (optional) The fields and values to filter by. Note these are exact matches.
      * @param IdInterface     $account         (optional) The account context to use in the request. Note that most requests require
      *                                         an account context.
      *
      * @return PromiseInterface A promise to return an ObjectList.
      */
-    public function selectRequest(ObjectListQuery $objectListQuery, IdInterface $account = null): PromiseInterface
+    public function selectRequest(ObjectListQuery $objectListQuery = null, IdInterface $account = null): PromiseInterface
     {
+        if (!$objectListQuery) {
+            $objectListQuery = new ObjectListQuery();
+        }
+
         $annotation = $this->dataService->getAnnotation();
         $options = [
             'query' => $objectListQuery->toQueryParts() + [