+++ tests/src/Functional/DataService/CustomFieldTest.php
@@ -2,10 +2,10 @@
 
 namespace Lullabot\Mpx\Tests\Functional\DataService;
 
-use Lullabot\Mpx\DataService\ObjectListQuery;
 use Lullabot\Mpx\DataService\DataObjectFactory;
 use Lullabot\Mpx\DataService\DataServiceManager;
 use Lullabot\Mpx\DataService\Field;
+use Lullabot\Mpx\DataService\ObjectListQuery;
 use Lullabot\Mpx\DataService\Range;
 use Lullabot\Mpx\Tests\Functional\FunctionalTestBase;