+++ tests/src/Functional/DataService/Media/MediaFileTest.php
@@ -5,7 +5,6 @@ namespace Lullabot\Mpx\Tests\Functional\DataService\Media;
 use Lullabot\Mpx\DataService\ObjectListQuery;
 use Lullabot\Mpx\DataService\DataObjectFactory;
 use Lullabot\Mpx\DataService\DataServiceManager;
-use Lullabot\Mpx\DataService\Range;
 use Lullabot\Mpx\Tests\Functional\FunctionalTestBase;
 use Psr\Http\Message\UriInterface;
 
@@ -23,10 +22,8 @@ class MediaFileTest extends FunctionalTestBase
         $service = $manager->getDataService('Media Data Service', 'MediaFile', '1.10');
         $dof = new DataObjectFactory($service, $this->authenticatedClient);
         $filter = new ObjectListQuery();
-        $range = new Range();
-        $range->setStartIndex(1)
+        $filter->getRange()->setStartIndex(1)
             ->setEndIndex(2);
-        $filter->setRange($range);
         $results = $dof->select($filter, $this->account);
 
         foreach ($results as $index => $result) {